Extracted AssemblyFinder from PluginManager to also be used in PluginSearcher. Modified DecendsFromOpenTapIgnore to expect a relative path.
