+++ Shared/PathUtils.cs
@@ -132,11 +132,18 @@ namespace Tap.Shared
             return Path.Combine(Path.GetTempPath(), Path.GetRandomFileName() + extension);
         }
 
+        /// <summary>
+        /// Checks if the relative path has any OpenTapIgnore in parent chain
+        /// </summary>
+        /// <param name="location">Relative path of file</param>
+        /// <returns>The precense of OpenTapIgnore in parent chain</returns>
         internal static bool DecendsFromOpenTapIgnore(string location)
         {
             string dir = Path.GetDirectoryName(location);
             if (File.Exists(Path.Combine(dir, ".OpenTapIgnore")))
                 return true;
+            if (string.IsNullOrWhiteSpace(dir))
+                return false;
             DirectoryInfo d = new DirectoryInfo(dir);
             if (d.Parent == null)
                 return false;
